refactor: current quiz component
