+++ src/components/index.js
@@ -1,6 +1,5 @@
 export { Navbar } from "./Navbar/Navbar";
 export { CategoryCard } from "./CategoryCard/CategoryCard";
 export { CurrentQuestion } from "./CurrentQuestion/CurrentQuestion";
-export { Score } from "./Score/Score";
 export { Loader } from "./Loader/Loader";
 export { Rules } from "./Rules/Rules";
